+++ packages/f-vue-icons/src/components/AmexSafekeyIcon/AmexSafekeyIcon.vue
@@ -1,0 +1,13 @@
+<template>
+    <amex-safekey-icon />
+</template>
+
+<script>
+import AmexSafekeyIcon from '@justeat/f-icons/src/img/icons/cards/amex--safekey.svg';
+
+export default {
+    components: {
+        AmexSafekeyIcon
+    }
+}
+</script>