+++ packages/f-vue-icons/src/components/InstagramIcon/InstagramIcon.vue
@@ -1,0 +1,13 @@
+<template>
+    <instagram-icon />
+</template>
+
+<script>
+import InstagramIcon from '@justeat/f-icons/src/img/icons/social/icon-instagram.svg';
+
+export default {
+    components: {
+        InstagramIcon
+    }
+}
+</script>