+++ packages/f-vue-icons/src/components/InteracIcon/InteracIcon.vue
@@ -1,0 +1,13 @@
+<template>
+    <interac-icon />
+</template>
+
+<script>
+import InteracIcon from '@justeat/f-icons/src/img/icons/cards/interac.svg';
+
+export default {
+    components: {
+        InteracIcon
+    }
+}
+</script>