+++ packages/f-vue-icons/src/components/InteracIcon/index.js
@@ -1,0 +1,7 @@
+import InteracIcon from './InteracIcon.vue';
+
+const install = Vue => {
+    Vue.component(InteracIcon.name, InteracIcon);
+};
+
+export default install;