+++ packages/f-vue-icons/src/components/MastercardSecurecodeIcon/MastercardSecurecodeIcon.vue
@@ -1,0 +1,13 @@
+<template>
+    <mastercard-securecode-icon />
+</template>
+
+<script>
+import MastercardSecurecodeIcon from '@justeat/f-icons/src/img/icons/cards/mastercard--securecode.svg';
+
+export default {
+    components: {
+        MastercardSecurecodeIcon
+    }
+}
+</script>