+++ packages/f-vue-icons/src/components/YouTubeIcon/YouTubeIcon.vue
@@ -1,0 +1,13 @@
+<template>
+    <youtube-icon />
+</template>
+
+<script>
+import YoutubeIcon from '@justeat/f-icons/src/img/icons/social/icon-youtube.svg';
+
+export default {
+    components: {
+        YoutubeIcon
+    }
+}
+</script>