+++ packages/f-vue-icons/CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.9.0
+------------------------------
+*July 10, 2019*
+
+### Added
+- Social icons
+- Payment/card icons
+
+
 v0.8.0
 ------------------------------
 *June 21, 2019*